wip: uberconf, appconf
