--- connector/src/uberconf/templates/rclone.conf
@@ -1,6 +1,0 @@
-[provider]
-type = webdav
-vendor = other
-url = {{WEBDAV_URL}}
-user = {{WEBDAV_USER}}
-pass = {{WEBDAV_PASS}}
