+++ connector/src/uberconf/uberconf.types.spec.ts
@@ -1,9 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  DebridConfig,
-  buildRcloneConf,
-  positiveDuration,
-} from "./uberconf.types";
+import { positiveDuration } from "./uberconf.types";
 
 describe("postiiveDuration", () => {
   it("validates as expected", () => {
@@ -25,22 +21,5 @@ describe("postiiveDuration", () => {
     expect(result3.error.issues[0].message).toContain(
       "Must be a positive duration string"
     );
-  });
-});
-
-describe("buildRcloneConf", () => {
-  it("builds the expected ini data", () => {
-    const config: DebridConfig = {
-      realDebrid: { username: "foo", password: "bar", apiKey: "baz" },
-    };
-    expect(buildRcloneConf(config)).toMatchInlineSnapshot(`
-      "[provider]
-      type = webdav
-      vendor = other
-      url = https://dav.real-debrid.com/
-      user = foo
-      pass = bar
-      "
-    `);
   });
 });
