+++ connector/src/uberconf/uberconf.types.ts
@@ -1,12 +1,6 @@
 import ms from "ms";
 import z from "zod";
-import { join } from "path";
-import { readFileSync, writeFileSync } from "fs";
-import ini from "ini";
 import { PROFILE_SCHEMA } from "../data/profile";
-
-const RCLONE_DUMMY_PASSWORD =
-  "m3PAV6DJIEGo4fuVlinAGtWZZXH0z_yabANkILj-ENwknWUnYkBDNK7TjbQ"; // rclone obscure dummy-password-please-ignore
 
 export const positiveDuration = z
   .string()
@@ -73,48 +67,3 @@ export const UBERCONF_SCHEMA = z.object({
   mediaProfiles: z.array(PROFILE_SCHEMA),
 });
 export type Uberconf = z.infer<typeof UBERCONF_SCHEMA>;
-
-function rcloneConf(dc: DebridConfig): {
-  WEBDAV_URL: string;
-  WEBDAV_USER: string;
-  WEBDAV_PASS: string;
-} {
-  if ("realDebrid" in dc) {
-    return {
-      WEBDAV_URL: "https://dav.real-debrid.com/",
-      WEBDAV_USER: dc.realDebrid.username,
-      WEBDAV_PASS: dc.realDebrid.password,
-    };
-  }
-  if ("allDebrid" in dc) {
-    return {
-      WEBDAV_URL: "https://webdav.debrid.it/",
-      WEBDAV_USER: dc.allDebrid.apiKey,
-      WEBDAV_PASS: RCLONE_DUMMY_PASSWORD,
-    };
-  }
-  const exhaustiveCheck: never = dc;
-  throw new Error(`unhandled debrid type: ${exhaustiveCheck}`);
-}
-
-export function buildRcloneConf(dc: DebridConfig): string {
-  const rc = rcloneConf(dc);
-  const data = {
-    provider: {
-      type: "webdav",
-      vendor: "other",
-      url: rc.WEBDAV_URL,
-      user: rc.WEBDAV_USER,
-      pass: rc.WEBDAV_PASS,
-    },
-  };
-  return ini.encode(data, { whitespace: true });
-}
-
-function writeConfigFiles(rootConfigDir: string, c: Uberconf) {
-  // writeTemplate(
-  //   "rclone.conf",
-  //   join(rootConfigDir, "rclone", "rclone.conf"),
-  //   rcloneConf(c)
-  // );
-}
